app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-keycloak>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiomysql>=0.2.0
cryptography>=43.0.0
alembic>=1.13.0
redis>=5.0.0
fastapi-limiter==0.1.6
PyJWT[crypto]>=2.9.0
httpx>=0.27.0
structlog>=24.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
